Extract GridLevel
