+++ include/components.hpp
@@ -4,7 +4,7 @@
 #include <glm/gtc/quaternion.hpp>
 #include <string>
 #include <vector>
-#include "Grid.hpp"
+#include "GridLevel.hpp"
 #include "Camera.hpp"
 #include "IntersectionSystem.hpp"
 #include "RobotDescription.hpp"
